requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
websockets
pydantic
pydantic-settings
pytest
requests
httpx
slowapi
